fix/pong-together/frontend#47: oauth 리다이렉트 url 수정
