+++ src/pages/components/login/Login.js
@@ -3,6 +3,7 @@ import OauthBtn from './OauthBtn.js';
 import SelectLanguage from './SelectLanguage.js';
 import TFABtn from './TFABtn.js';
 import store from '../../../store/index.js';
+import { navigate } from '../../../router/utils/navigate.js';
 
 export default class extends Component {
 	setup() {
@@ -23,8 +24,8 @@ export default class extends Component {
 		const $parent = this.$target.querySelector('.login-body-wrapper');
 
 		if (store.state.loginProgress === 'done') {
-			//location.pathname = '/';
 			window.location.pathname = '/select';
+			//navigate('/select');
 		}
 		if (store.state.loginProgress === 'oauth') {
 			new OauthBtn($parent);
