+++ src/router/constants/pages.js
@@ -4,12 +4,14 @@ import Tournament from '../../pages/components/tournament/Tournament.js';
 import Remote from '../../pages/components/Remote/Remote.js';
 import Local from '../../pages/components/Local.js';
 import Game from '../../pages/components/game/Game.js';
+import LoginRedirect from '../../pages/components/login/LoginRedirect';
 
-export const BASE_URL = 'https://localhost:3000';
+export const BASE_URL = 'https://localhost:443';
 
 export const routes = [
 	{ path: /^\/$/, element: Login },
 	{ path: /^\/login$/, element: Login },
+	{ path: /^\/auth$/, element: LoginRedirect },
 	{ path: /^\/select$/, element: GameSelect },
 	{ path: /^\/local$/, element: Local },
 	{ path: /^\/tournament$/, element: Tournament },
